+++ CHANGES.md
@@ -76,6 +76,12 @@
   let f x = x.Unix.tms_stime, x.Unix.tms_utime
   ```
 
+- Bug fix: do not show "random" documentation on hover
+  - fixed by [merlin#1364](https://github.com/ocaml/merlin/pull/1364)
+  - fixes duplicate:
+    - [ocaml-lsp#344](https://github.com/ocaml/ocaml-lsp/issues/344)
+    - [vscode-ocaml-platform#111](https://github.com/ocamllabs/vscode-ocaml-platform/issues/111)
+
 # 1.7.0 (07/28/2021)
 
 ## Features